requirements.txt
mysql-connector-python
PyMySQL